change center and obj id to i32

+++ src/neospy/rust/fovs/checks.rs
@@ -64,7 +64,7 @@ pub fn fov_checks_py(
 
 #[pyfunction]
 #[pyo3(name = "fov_spk_checks")]
-pub fn fov_spk_checks_py(obj_ids: Vec<isize>, fovs: FOVListLike) -> Vec<PySimultaneousStates> {
+pub fn fov_spk_checks_py(obj_ids: Vec<i32>, fovs: FOVListLike) -> Vec<PySimultaneousStates> {
     let fovs = fovs.into_sorted_vec_fov();
 
     fovs.into_par_iter()
